backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3

# Document processing
pdfplumber==0.10.3
python-docx==1.1.0
pypandoc==1.12

# LangChain and AI
langchain==0.1.4
langchain-google-genai==0.0.6
langchain-community==0.0.16
chromadb==0.4.22
onnxruntime==1.17.0

# Report generation
reportlab==4.0.9

# Utilities
python-dotenv==1.0.0
Pillow==10.2.0
